+++ src/Daedalus.jl
@@ -13,9 +13,10 @@ include("Ensemble.jl")
 include("DaedalusOutput.jl")
 include("Outputs.jl")
 include("Costs.jl")
+include("NpiHelpers.jl")
 
-export Npi, ParamEffect, ReactiveTrigger, TimeTrigger, get_values, get_times,
-       DaedalusOutput, get_data, get_incidence, get_epidemic_summary, get_life_years_lost,
-       get_costs
+export Npi, Vaccination, ParamEffect, ReactiveTrigger, TimeTrigger, get_values, get_times,
+       DaedalusOutput, get_data, get_npi, get_vaccination, get_incidence, get_epidemic_summary, get_life_years_lost,
+       get_costs, create_time_intervention, create_reactive_intervention, create_multi_intervention
 
 end